backend/requirements.txt
# FastAPI 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# RAG 框架
llama-index>=0.9.0
langchain>=0.3.3
langchain-text-splitters>=1.0.0

# 向量数据库
faiss-cpu>=1.7.4
# 如果有 GPU，可以使用 faiss-gpu
# faiss-gpu>=1.7.4

#
rank-bm25==0.2.2

# 文档解析
python-docx>=1.1.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# 重排模型（Jina Reranker API）
requests>=2.31.0

# Qwen API 客户端
#openai>=1.3.0
# 或者使用 dashscope（阿里云官方 SDK）
dashscope>=1.17.0

# 环境变量管理
python-dotenv>=1.0.0

# 其他工具
numpy>=1.24.0
PyPDF2==3.0.1
pandas==2.2.3
tiktoken>=0.8.0
tqdm>=4.66.0
jieba>=0.42.1

